# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ViSynth API
  version: 0.3.0
  description: |
    API REST của ViSynth (tên tạm). Gốc: /api/v1. Xác thực bằng cookie phiên HttpOnly (web) hoặc Bearer token (giai đoạn 2).
    Quy ước:
    - Lỗi theo RFC 9457 (application/problem+json) kèm trường `code` ổn định để client xử lý.
    - POST /jobs bắt buộc có header `Idempotency-Key` (chống tạo trùng job và trừ tín dụng hai lần).
    - Tiến độ job qua SSE: GET /jobs/{jobId}/events (hỗ trợ Last-Event-ID).
    - Hợp đồng dữ liệu chi tiết của prompt/pipeline nằm trong thư mục schemas/.
servers:
- {url: /api/v1}
security:
- cookieAuth: []
- bearerAuth: []
...
paths:
...
  /admin/pool/config:
    get:
      tags: [pool]
      operationId: adminExportPoolConfig
      summary: Xuất PoolConfig (schemas/pool_config.schema.json). Khoá chỉ ở dạng tham chiếu secret_ref, không bao giờ trả khoá thật
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {type: object, description: PoolConfig}
        '403': {$ref: '#/components/responses/Forbidden'}
    put:
      tags: [pool]
      operationId: adminImportPoolConfig
      summary: Nhập PoolConfig. dry_run=true (mặc định) chỉ kiểm tra schema, ràng buộc chéo và trả bản so sánh thay đổi; dry_run=false áp dụng và tăng app_settings.pool_version
      parameters:
      - name: dry_run
        in: query
        schema: {type: boolean, default: true}
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, description: PoolConfig}
      responses:
        '200':
          description: Kết quả kiểm tra/áp dụng
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PoolImportResult'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '422': {$ref: '#/components/responses/Problem'}
  /admin/pool/status:
    get:
      tags: [pool]
      operationId: adminPoolStatus
      summary: 'Trạng thái sống của từng deployment: headroom, circuit, cooldown, đã dùng trong ngày, số khoá đang hoạt động'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/PoolDeploymentStatus'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /admin/pool/capacity:
    get:
      tags: [pool]
      operationId: adminPoolCapacity
      summary: 'Dự báo dung lượng: số tài liệu mỗi ngày theo mức và theo chế độ riêng tư, tính từ hạn mức đã cấu hình (reference/estimator.py: estimate_calls, docs_per_day)'
      parameters:
      - name: privacy_class
        in: query
        schema: {$ref: '#/components/schemas/PrivacyClass'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PoolCapacity'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /admin/pool/groups/{groupId}:
    parameters:
    - name: groupId
      in: path
      required: true
      schema: {type: string}
    patch:
      tags: [pool]
      operationId: adminPatchPoolGroup
      summary: 'Sửa nhóm hạn mức: bật/tắt, hạn mức cả tài khoản, hệ số an toàn, cờ ToS, cổng triển khai được phép, data_policy (ghi audit_log)'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PoolGroupPatch'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PoolGroup'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /admin/pool/groups/{groupId}/credentials:
    parameters:
    - name: groupId
      in: path
      required: true
      schema: {type: string}
    post:
      tags: [pool]
      operationId: adminAddPoolCredential
      summary: Thêm một khoá vào nhóm. secret chỉ ghi, mã hoá ở tầng ứng dụng (secret_enc); phản hồi chỉ có 4 ký tự cuối
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PoolCredentialCreate'}
      responses:
        '201':
          description: Đã thêm
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PoolCredential'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Problem'}
  /admin/pool/credentials/{credentialId}:
    parameters:
    - name: credentialId
      in: path
      required: true
      schema: {type: string}
    patch:
      tags: [pool]
      operationId: adminPatchPoolCredential
      summary: Bật/tắt khoá hoặc gỡ trạng thái quarantined sau khi đã xử lý nguyên nhân bị từ chối
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [active, disabled]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PoolCredential'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [pool]
      operationId: adminDeletePoolCredential
      summary: Xoá khoá (xoá bản mã hoá khỏi CSDL)
      responses:
        '204': {description: Đã xoá}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /admin/pool/deployments/{deploymentId}:
    parameters:
    - name: deploymentId
      in: path
      required: true
      schema: {type: string}
    patch:
      tags: [pool]
      operationId: adminPatchPoolDeployment
      summary: Sửa hạn mức, trọng số, thẻ, bật/tắt một deployment
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PoolDeploymentPatch'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PoolDeploymentStatus'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /admin/pool/deployments/{deploymentId}/probe:
    parameters:
    - name: deploymentId
      in: path
      required: true
      schema: {type: string}
    post:
      tags: [pool]
      operationId: adminProbeDeployment
      summary: 'Chạy bài kiểm định nhận vào pool: tuân thủ JSON theo schema, tiếng Việt, ngữ cảnh dài, dịch máy; ghi llm_probe_runs và cập nhật quality của model'
      responses:
        '202':
          description: Đã nhận
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProbeRun'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /admin/pool/incidents:
    get:
      tags: [pool]
      operationId: adminPoolIncidents
      summary: 'Sự cố của pool: 429, mở circuit, khoá bị từ chối'
      parameters:
      - {$ref: '#/components/parameters/Cursor'}
      - {$ref: '#/components/parameters/Limit'}
      - name: deployment_id
        in: query
        schema: {type: string}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/PoolIncident'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /admin/pool/declare:
    post:
      tags: [pool]
      operationId: adminDeclarePool
      summary: 'Khai báo nhà cung cấp và khoá hàng loạt (phương pháp nhập thông tin, SPEC §17.14). Body: PoolDeclaration (JSON, hoặc YAML với Content-Type application/yaml); khoá cách nhau bằng dấu phẩy, chấm phẩy, xuống dòng, có thể kèm nhãn nhãn|khoá. dry_run=true (mặc định) chỉ biên dịch và trả bản xem trước, KHÔNG ghi gì; dry_run=false ghi trong một giao dịch, mã hoá khoá ngay và chỉ trả 4 ký tự cuối. Endpoint này không được ghi log nội dung yêu cầu'
      parameters:
      - name: dry_run
        in: query
        schema: {type: boolean, default: true}
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, description: PoolDeclaration - schemas/pool_declaration.schema.json}
          application/yaml:
            schema: {type: string, description: 'Cùng nội dung ở dạng YAML (mẫu: examples/pool_declaration.template.yaml)'}
      responses:
        '200':
          description: Kết quả biên dịch hoặc áp dụng
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DeclareResult'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Problem'}
  /admin/pool/declaration-template:
    get:
      tags: [pool]
      operationId: adminPoolDeclarationTemplate
      summary: Mẫu khai báo YAML có chú thích bằng tiếng Việt để điền (cùng nội dung examples/pool_declaration.template.yaml)
      responses:
        '200':
          description: OK
          content:
            application/yaml:
              schema: {type: string}
        '403': {$ref: '#/components/responses/Forbidden'}
components:
  securitySchemes:
    cookieAuth: {type: apiKey, in: cookie, name: visynth_session}
    bearerAuth: {type: http, scheme: bearer}
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      schema: {type: string}
      description: Con trỏ phân trang không trong suốt
    Limit:
      name: limit
      in: query
      required: false
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
  responses:
    Problem:
      description: Lỗi
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
    Forbidden:
      description: Không đủ quyền
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    NotFound:
      description: Không tìm thấy (cũng trả về khi tài nguyên thuộc người dùng khác - chống dò ID)
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Conflict:
      description: Xung đột trạng thái
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: {type: string}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        code:
          type: string
          enum: [unauthenticated, forbidden, not_found, validation_error, unsupported_file_type, file_too_large, document_too_large, extraction_failed, encrypted_pdf, insufficient_credits, rate_limited, spend_cap_reached, invite_invalid, invite_exhausted, invite_already_redeemed, job_not_cancelable, glossary_not_ready, rights_not_attested, consent_required, active_job_limit, document_expired, internal_error, level_disabled, full_translation_daily_limit, shared_processing_consent_required, age_confirmation_required, privacy_mode_unavailable, pool_capacity_exceeded, approval_blocked, version_immutable, style_core_not_approved, draft_not_editable, risk_ack_required, declaration_invalid]
        errors:
          type: array
          items:
            type: object
            properties:
              field: {type: string}
              message: {type: string}
    Level:
      type: string
      enum: [full_translation, detailed_synthesis, deep_synthesis, executive_brief]
...
    PrivacyClass:
      type: string
      enum: [standard, private]
    GroupTier:
      type: string
      enum: [free, trial, paid, self_hosted]
    DataPolicy:
      type: string
      enum: [no_training, may_train, unknown]
...
    PoolLimits:
      type: object
      properties:
        rpm:
          type: [integer, 'null']
          minimum: 1
        tpm:
          type: [integer, 'null']
          minimum: 1
        rpd:
          type: [integer, 'null']
          minimum: 1
        tpd:
          type: [integer, 'null']
          minimum: 1
        concurrency:
          type: [integer, 'null']
          minimum: 1
    PoolGroup:
      type: object
      required: [id, provider, tier, data_policy, enabled]
      properties:
        id: {type: string}
        provider: {type: string}
        label: {type: string}
        tier: {$ref: '#/components/schemas/GroupTier'}
        data_policy: {$ref: '#/components/schemas/DataPolicy'}
        tos_flags:
          type: array
          items:
            type: string
            enum: [trial_only, no_personal_data, multi_account_risk, no_eea_uk_ch]
        allowed_gates:
          type: array
          items:
            type: string
            enum: [dev, A, B, C]
        safety_margin: {type: number}
        day_margin: {type: number}
        limits: {$ref: '#/components/schemas/PoolLimits'}
        enabled: {type: boolean}
        risk_ack:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [multi_account_risk, trial_only]
          description: Cờ rủi ro điều khoản đã được chủ hệ thống xác nhận chấp nhận. Nhóm có cờ tương ứng trong tos_flags mà thiếu xác nhận không được bật (409 risk_ack_required) và không bao giờ được chọn.
    PoolGroupPatch:
      type: object
      properties:
        enabled: {type: boolean}
        data_policy: {$ref: '#/components/schemas/DataPolicy'}
        tos_flags:
          type: array
          items:
            type: string
            enum: [trial_only, no_personal_data, multi_account_risk, no_eea_uk_ch]
        allowed_gates:
          type: array
          minItems: 1
          items:
            type: string
            enum: [dev, A, B, C]
        safety_margin: {type: number, minimum: 0.3, maximum: 1}
        day_margin: {type: number, minimum: 0.5, maximum: 1}
        limits: {$ref: '#/components/schemas/PoolLimits'}
        risk_ack:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [multi_account_risk, trial_only]
          description: Cờ rủi ro điều khoản đã được chủ hệ thống xác nhận chấp nhận. Nhóm có cờ tương ứng trong tos_flags mà thiếu xác nhận không được bật (409 risk_ack_required) và không bao giờ được chọn.
    PoolDeploymentPatch:
      type: object
      properties:
        enabled: {type: boolean}
        weight: {type: number, exclusiveMinimum: 0}
        tags:
          type: array
          items: {type: string}
        limits: {$ref: '#/components/schemas/PoolLimits'}
    PoolDeploymentStatus:
      type: object
      required: [id, group_id, model, tier, enabled, circuit, headroom]
      properties:
        id: {type: string}
        group_id: {type: string}
        model: {type: string}
        tier: {$ref: '#/components/schemas/GroupTier'}
        data_policy: {$ref: '#/components/schemas/DataPolicy'}
        enabled: {type: boolean}
        circuit:
          type: string
          enum: [closed, open, half_open]
        cooldown_until:
          type: [string, 'null']
          format: date-time
        headroom: {type: number, minimum: 0, maximum: 1}
        inflight: {type: integer}
        rpd_used: {type: integer}
        rpd_limit:
          type: [integer, 'null']
        ewma_success: {type: number}
        ewma_latency_ms:
          type: [number, 'null']
        active_credentials: {type: integer}
        limits: {$ref: '#/components/schemas/PoolLimits'}
    PoolCredential:
      type: object
      required: [id, group_id, label, status]
      properties:
        id: {type: string}
        group_id: {type: string}
        label: {type: string}
        last4:
          type: [string, 'null']
        status:
          type: string
          enum: [active, quarantined, disabled]
        quarantined_reason:
          type: [string, 'null']
        last_used_at:
          type: [string, 'null']
          format: date-time
    PoolCredentialCreate:
      type: object
      required: [label, secret]
      properties:
        label: {type: string, maxLength: 120}
        secret: {type: string, minLength: 16, maxLength: 400, writeOnly: true}
    PoolCapacity:
      type: object
      required: [privacy_class, eligible_deployments, docs_per_day]
      properties:
        privacy_class: {$ref: '#/components/schemas/PrivacyClass'}
        eligible_deployments: {type: integer}
        calls_per_day:
          type: [integer, 'null']
        tokens_in_per_day:
          type: [integer, 'null']
        docs_per_day:
          type: array
          items:
            type: object
            required: [level, words, docs]
            properties:
              level: {$ref: '#/components/schemas/Level'}
              words: {type: integer}
              docs:
                type: [number, 'null']
                description: null = không có giới hạn đã biết
    PoolImportResult:
      type: object
      required: [valid, applied, errors, diff]
      properties:
        valid: {type: boolean}
        applied: {type: boolean}
        errors:
          type: array
          items: {type: string}
        warnings:
          type: array
          description: Cảnh báo không chặn (ví dụ trường chỉ để hiển thị bị bỏ qua khi nhập lại bản xuất).
          items: {type: string}
        version:
          type: integer
          description: Số phiên bản cấu hình pool sau khi ghi (`app_settings.pool_version`).
        diff:
          type: object
          properties:
            added:
              type: array
              items: {type: string}
            removed:
              type: array
              items: {type: string}
            changed:
              type: array
              items: {type: string}
    ProbeRun:
      type: object
      required: [id, deployment_id, passed]
      properties:
        id: {type: integer}
        deployment_id: {type: string}
        ran_at: {type: string, format: date-time}
        passed: {type: boolean}
        results: {type: object}
        note:
          type: [string, 'null']
    PoolIncident:
      type: object
      required: [id, at, kind]
      properties:
        id: {type: integer}
        at: {type: string, format: date-time}
        deployment_id:
          type: [string, 'null']
        credential_id:
          type: [string, 'null']
        kind: {type: string}
        detail: {type: object}
...
    DeclareResult:
      type: object
      required: [valid, applied, errors, warnings, preview]
      properties:
        valid: {type: boolean}
        applied: {type: boolean}
        errors:
          type: array
          items: {type: string}
        warnings:
          type: array
          items: {type: string}
        preview:
          type: object
          properties:
            counts:
              type: object
              properties:
                groups: {type: integer}
                new_groups: {type: integer}
                keys: {type: integer}
                skipped_keys: {type: integer}
                deployments: {type: integer}
                models: {type: integer}
                providers: {type: integer}
            groups:
              type: array
              items:
                type: object
                properties:
                  id: {type: string}
                  existing: {type: boolean}
                  tier: {$ref: '#/components/schemas/GroupTier'}
                  data_policy: {$ref: '#/components/schemas/DataPolicy'}
                  tos_flags:
                    type: array
                    items: {type: string}
                  allowed_gates:
                    type: array
                    items: {type: string}
                  risk_ack:
                    type: array
                    items: {type: string}
                  account_limits:
                    type: [object, 'null']
                  credentials:
                    type: array
                    items:
                      type: object
                      properties:
                        id: {type: string}
                        last4: {type: string}
                        status: {type: string}
                  deployments:
                    type: array
                    items:
                      type: object
                      properties:
                        id: {type: string}
                        model: {type: string}
                        limits: {$ref: '#/components/schemas/PoolLimits'}
                        limits_source: {type: string}
            skipped_keys:
              type: array
              description: Khoá bị bỏ qua và lý do (trùng, quá ngắn, sai định dạng, giữ chỗ). Không bao giờ chứa khoá đầy đủ.
              items:
                type: object
                properties:
                  label:
                    type: [string, 'null']
                  last4: {type: string}
                  status:
                    type: string
                    enum: [duplicate_in_request, duplicate_existing, too_short, invalid, placeholder]
                  reason: {type: string}
            validate_keys:
              type: string
              enum: [pending, ok, failed]